tests/wpt/tests/workers/interfaces/WorkerUtils/importScripts/003.html

<!doctype html>
<title>importScripts running scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('003.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, "abc");
    this.done();
  });
});
</script>
